move `remove_nesting_layer` to `utils`
